requirements.txt
Appium-Python-Client>=3.2.1
pytest>=8.0.0
pytest-html>=4.1.1
selenium>=4.18.1
nltk>=3.8.1
numpy>=1.26.4
pandas>=2.2.1
matplotlib>=3.8.3
python-dotenv>=1.0.1
webdriver-manager>=4.0.1